# docs/openapi.yaml
openapi: 3.0.0
info:
  title: crPipeline API
  version: "1.0.0"
servers:
  - url: http://localhost:8080/api
paths:
...
  /upload:
    post:
      summary: Upload document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Document uploaded
  /download/{document_id}:
    get:
      summary: Download a document
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document content or presigned URL
...